actions and live numbering with sorts
- expected still missing
- also changes start symbol to always be just "S" not "#S"

+++ common/src/table/state.rs
@@ -3,4 +3,4 @@
 
 // }
 
-pub type StateId = u32;
+pub type StateId = usize;
